initialised tests, nuke functions and custom query's

+++ android/src/com/mygdx/game/persistance/Coordinate/NNCoordinate.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity(tableName = "coordinate")
 public class NNCoordinate {
     @PrimaryKey(autoGenerate = true)
-    public int id;
-    public int x;
-    public int y;
+    public long  id;
+    public int  x;
+    public int  y;
 }
